Allow customizable Standard wish pools and 10-pull cost
Allow for users to change the 10-pull cost, and to customize Standard wish pulls on banners.

Note: While the discount WILL show up in-game, it will only do so before making a wish. It will revert to showing a flat 10-cost afterward, but will still only require/consume the desired amount for a 10-pull.