feature(support 2.8): upgrade grasscutter to 2.8
1. Upload new protos, and I added proto for almost all activities to help developers implement more activities. To solve the memory issues, add `gradle.properties`;
2. Fix some bugs cuz by new protos, a painful experience that includes a lot of guesswork;
3. Modified the login logic, from now on if you want to use the grasscutter for this branch you have to patch your `UserAssembly.dll`, but keep the seed exchange logic.

Because of the obfuscation, a large number of proto field names have been changed, and change to the original de-obfuscated logic and fields is unavoidable, so I can only rely on my personal guesses and related logical to fix them. I apologize if the functionality is broken.

BREAKING CHANGE: ALL